Fixed methods

+++ demo/src/main/java/com/epam/demo/Utils.java
@@ -5,6 +5,6 @@ import org.apache.commons.lang3.math.NumberUtils;
 
 public class Utils {
     public static boolean isAllPositiveNumbers(List<String> args) {
-        return args.stream().allMatch(x -> NumberUtils.isCreatable(x) && NumberUtils.toInt(x)>0);
+        return args.stream().allMatch(x -> NumberUtils.isCreatable(x) && NumberUtils.toDouble(x)>0);
     }
 }
